Moved all error handling to checkProducts and checkPromotions, tested error handling
