+++ db/menu.db
@@ -1,1 +1,1 @@
-{"type":"menu","data":[{"_id":1,"title":"Bryggkaffe","desc":"Bryggd på månadens bönor.","price":39,"createdAt":"07/06/2024, 22:50"},{"_id":2,"title":"Caffè Doppio","desc":"Bryggd på månadens bönor.","price":49,"createdAt":"07/06/2024, 22:50"},{"_id":3,"title":"Cappuccino","desc":"Bryggd på månadens bönor.","price":49,"createdAt":"07/06/2024, 22:50"},{"_id":4,"title":"Latte Macchiato","desc":"Bryggd på månadens bönor.","price":49,"createdAt":"07/06/2024, 22:50"},{"_id":5,"title":"Kaffe Latte","desc":"Bryggd på månadens bönor.","price":54,"createdAt":"07/06/2024, 22:50"},{"_id":6,"title":"Cortado","desc":"Bryggd på månadens bönor.","price":39,"createdAt":"07/06/2024, 22:50"}],"_id":"uPqjz2Yp8U3Jakge"}
+{"type":"menu","data":[{"_id":1,"title":"Bryggkaffe","desc":"Bryggd på månadens bönor.","price":39,"createdAt":"10/06/2024, 14:35"},{"_id":2,"title":"Caffè Doppio","desc":"Bryggd på månadens bönor.","price":49,"createdAt":"10/06/2024, 14:35"},{"_id":3,"title":"Cappuccino","desc":"Bryggd på månadens bönor.","price":49,"createdAt":"10/06/2024, 14:35"},{"_id":4,"title":"Latte Macchiato","desc":"Bryggd på månadens bönor.","price":49,"createdAt":"10/06/2024, 14:35"},{"_id":5,"title":"Kaffe Latte","desc":"Bryggd på månadens bönor.","price":54,"createdAt":"10/06/2024, 14:35"},{"_id":6,"title":"Cortado","desc":"Bryggd på månadens bönor.","price":39,"createdAt":"10/06/2024, 14:35"}],"_id":"ZVqt6MwqvhUDRagx"}
